simplify backend macros and TS
