+++ backend/macros.ts
@@ -4,9 +4,8 @@
  */
 
 import path from 'path';
-import URI from 'urijs';
 import fs from 'fs-extra';
-import rollbar from 'rollbar';
+import Rollbar, { MaybeError } from 'rollbar';
 import Amplitude from 'amplitude';
 
 import commonMacros from '../common/abstractMacros';
@@ -22,8 +21,8 @@ const amplitude = new Amplitude(commonMacros.amplitudeToken);
 // TODO: improve getBaseHost by using a list of top level domains. (public on the internet)
 
 // Change the current working directory to the directory with package.json and .git folder.
-const originalCwd = process.cwd();
-let oldcwd;
+const originalCwd: string = process.cwd();
+let oldcwd: string;
 while (1) {
   try {
     fs.statSync('package.json');
@@ -45,7 +44,6 @@ while (1) {
   break;
 }
 
-
 // This is the JSON object saved in /etc/searchneu/config.json
 // null = hasen't been loaded yet.
 // {} = it has been loaded, but nothing was found or the file doesn't exist or the file was {}
@@ -53,143 +51,22 @@ while (1) {
 let envVariables = null;
 
 class Macros extends commonMacros {
-  static parseNameWithSpaces(name) {
-    // Standardize spaces.
-    name = name.trim().replace(/\s+/gi, ' ');
+// Version of the schema for the data. Any changes in this schema will effect the data saved in the dev_data folder
+  // and the data saved in the term dumps in the public folder and the search indexes in the public folder.
+  // Increment this number every time there is a breaking change in the schema.
+  // This will cause the data to be saved in a different folder in the public data folder.
+  // The first schema change is here: https://github.com/ryanhugh/searchneu/pull/48
+  static schemaVersion = 2;
 
-    // Generate first name and last name
-    const spaceCount = Macros.occurrences(name, ' ', false);
-    const splitName = name.split(' ');
+  static PUBLIC_DIR = path.join('public', 'data', `v${Macros.schemaVersion}`);
 
+  static DEV_DATA_DIR = path.join('dev_data', `v${Macros.schemaVersion}`);
 
-    if (spaceCount === 0) {
-      Macros.warn('0 spaces found in name', name);
-      return null;
-    }
+  // Folder of the raw html cache for the requests.
+  static REQUESTS_CACHE_DIR = 'requests';
 
-    // Handles firstName, lastName and firstName, middleName, lastName
-
-    if (spaceCount > 2) {
-      Macros.log(name, 'has more than 1 space in their name. Using first and last word.');
-    }
-
-    const obj = {};
-
-    obj.firstName = splitName[0];
-    obj.lastName = splitName[splitName.length - 1];
-
-    return obj;
-  }
-
-  // Standardizes email addresses found across different pages
-  // Removes a 'mailto:' from the beginning
-  // Ensures the email contains a @
-  static standardizeEmail(email) {
-    if (!email) {
-      return null;
-    }
-
-    if (email.startsWith('mailto:')) {
-      email = email.slice('mailto:'.length);
-    }
-
-    if (!email.includes('@') || email.includes(' ')) {
-      return null;
-    }
-
-    if (email.endsWith('@neu.edu')) {
-      email = `${email.split('@')[0]}@northeastern.edu`;
-    }
-
-    return email.toLowerCase().trim();
-  }
-
-
-  static standardizePhone(phone) {
-    if (!phone) {
-      return null;
-    }
-
-    phone = phone.trim();
-
-    if (phone.startsWith('tel:')) {
-      phone = phone.slice('tel:'.length).trim();
-    }
-
-    let digitsOnly = phone.replace(/[^0-9]/gi, '');
-
-
-    if (phone.startsWith('+1') && digitsOnly.length === 11) {
-      digitsOnly = digitsOnly.slice(1);
-    }
-
-    if (digitsOnly.length !== 10) {
-      return null;
-    }
-
-    return digitsOnly;
-  }
-
-  // Parses the google scholar id from a link that should contain a google scholar link.
-  // Get the Google Scholar ID with this: https://scholar.google.com/citations?user=[id here]
-  static parseGoogleScolarLink(link) {
-    if (!link) {
-      return null;
-    }
-
-    const userId = new URI(link).query(true).user;
-    if (!userId && link) {
-      Macros.log('Error parsing google url', link);
-      return null;
-    }
-    return userId;
-  }
-
-
-  // Gets the base hostname from a url.
-  // fafjl.google.com -> google.com
-  // subdomain.bob.co -> bob.co
-  // bob.co -> bob.co
-  // This could be improved by using public lists of top-level domains.
-  static getBaseHost(url) {
-    const homepage = new URI(url).hostname();
-    if (!homepage || homepage === '') {
-      Macros.error('could not find homepage of', url);
-      return null;
-    }
-
-    const match = homepage.match(/[^.]+\.[^.]+$/i);
-    if (!match) {
-      Macros.error('homepage match failed...', homepage);
-      return null;
-    }
-    return match[0];
-  }
-
-
-  // http://stackoverflow.com/questions/4009756/how-to-count-string-occurrence-in-string/7924240#7924240
-  static occurrences(string, subString, allowOverlapping) {
-    string += '';
-    subString += '';
-    if (subString.length <= 0) {
-      return (string.length + 1);
-    }
-
-    let n = 0;
-    let pos = 0;
-    const step = allowOverlapping ? 1 : subString.length;
-
-    while (true) {
-      pos = string.indexOf(subString, pos);
-      if (pos >= 0) {
-        ++n;
-        pos += step;
-      } else {
-        break;
-      }
-    }
-    return n;
-  }
+  // For iterating over every letter in a couple different places in the code.
+  static ALPHABET = 'maqwertyuiopsdfghjklzxcvbn';
 
   static getAllEnvVariables() {
     if (envVariables) {
@@ -224,7 +101,7 @@ class Macros extends commonMacros {
   }
 
   // Log an event to amplitude. Same function signature as the function for the frontend.
-  static async logAmplitudeEvent(type, event) {
+  static async logAmplitudeEvent(type: string, event: any) {
     if (!Macros.PROD) {
       return null;
     }
@@ -244,7 +121,7 @@ class Macros extends commonMacros {
   // Takes an array of a bunch of thigs to log to rollbar
   // Any of the times in the args array can be an error, and it will be logs according to rollbar's API
   // shouldExit - exit after logging.
-  static async logRollbarError(args, shouldExit) {
+  static async logRollbarError(args: any, shouldExit: boolean) {
     // Don't log rollbar stuff outside of Prod
     if (!Macros.PROD) {
       return;
@@ -258,7 +135,7 @@ class Macros extends commonMacros {
       return;
     }
 
-    rollbar.init(rollbarKey);
+    const rollbar = Rollbar.init(rollbarKey);
 
     const stack = (new Error()).stack;
 
@@ -266,7 +143,7 @@ class Macros extends commonMacros {
     args.stack = stack;
 
     // Search through the args array for an error. If one is found, log that separately.
-    let possibleError;
+    let possibleError: MaybeError;
 
     for (const value of Object.values(args)) {
       if (value instanceof Error) {
@@ -297,7 +174,7 @@ class Macros extends commonMacros {
 
   // This is for programming errors. This will cause the program to exit anywhere.
   // This *should* never be called.
-  static critical(...args) {
+  static critical(...args: any) {
     if (Macros.TESTS) {
       console.error('macros.critical called'); // eslint-disable-line no-console
       console.error(...args); // eslint-disable-line no-console
@@ -310,7 +187,7 @@ class Macros extends commonMacros {
 
   // Use this for stuff that is bad, and shouldn't happen, but isn't mission critical and can be ignored and the app will continue working
   // Will log something to rollbar and rollbar will send off an email
-  static async warn(...args) {
+  static async warn(...args: any) {
     super.warn(...args);
 
     if (Macros.PROD) {
@@ -324,7 +201,7 @@ class Macros extends commonMacros {
   // Will log stack trace
   // and cause CI to fail
   // so CI will send an email
-  static async error(...args) {
+  static async error(...args: any) {
     super.error(...args);
 
     if (Macros.PROD) {
@@ -340,7 +217,7 @@ class Macros extends commonMacros {
   }
 
   // Use console.warn to log stuff during testing
-  static verbose(...args) {
+  static verbose(...args: any) {
     if (!process.env.VERBOSE) {
       return;
     }
@@ -349,36 +226,19 @@ class Macros extends commonMacros {
   }
 }
 
-// Version of the schema for the data. Any changes in this schema will effect the data saved in the dev_data folder
-// and the data saved in the term dumps in the public folder and the search indexes in the public folder.
-// Increment this number every time there is a breaking change in the schema.
-// This will cause the data to be saved in a different folder in the public data folder.
-// The first schema change is here: https://github.com/ryanhugh/searchneu/pull/48
-Macros.schemaVersion = 2;
-
-Macros.PUBLIC_DIR = path.join('public', 'data', `v${Macros.schemaVersion}`);
-Macros.DEV_DATA_DIR = path.join('dev_data', `v${Macros.schemaVersion}`);
-
-// Folder of the raw html cache for the requests.
-Macros.REQUESTS_CACHE_DIR = 'requests';
-
-// For iterating over every letter in a couple different places in the code.
-Macros.ALPHABET = 'maqwertyuiopsdfghjklzxcvbn';
-
 
 Macros.verbose('Starting in verbose mode.');
 
 
-async function handleUncaught(err) {
+async function handleUncaught(err: Error) {
   // Don't use the macros.log, because if that crashes it would run into an infinite loop
   console.log('Error: An unhandledRejection occurred.'); // eslint-disable-line no-console
   console.log(`Rejection Stack Trace: ${err.stack}`); // eslint-disable-line no-console
   Macros.error(err.stack);
 }
 
-
 // Sometimes it helps debugging to enable this test mode too.
-if ((Macros.PROD || Macros.DEV || 1) && !global.addedRejectionHandler) {
+if (Macros.PROD && !global.addedRejectionHandler) {
   global.addedRejectionHandler = true;
   process.on('unhandledRejection', handleUncaught);
   process.on('uncaughtException', handleUncaught);
