+++ package.json
@@ -92,7 +92,7 @@
     "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
     "@types/react-router-dom": "^5.1.3",
-    "amplitude": "^4.0.0",
+    "amplitude": "^4.0.1",
     "apollo-server": "^2.9.5",
     "apollo-server-express": "^2.9.7",
     "apollo-server-testing": "^2.9.6",
